requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
python-telegram-bot[webhooks]>=22.6
agno>=2.5.7
google-genai>=0.5.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
google-cloud-firestore>=2.15.0
